Check if all values in a groupby column are NaN or not NaN
